Import precheck: Refactor to allow mocking
