+++ cli_tools/import_precheck/main.go
@@ -1,15 +1,17 @@
-/*
-Copyright 2017 Google Inc. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
 package main
 
 import (
@@ -17,6 +19,8 @@ import (
 	"io"
 	"os"
 	"sync"
+
+	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/import_precheck/precheck"
 
 	"github.com/GoogleCloudPlatform/osconfig/osinfo"
 	"github.com/google/logger"
@@ -28,13 +32,13 @@ var (
 	log *logger.Logger
 )
 
-func getChecks(osInfo *osinfo.OSInfo) []check {
-	return []check{
-		&osVersionCheck{osInfo},
-		&disksCheck{},
-		&sshCheck{},
-		&powershellCheck{},
-		&sha2DriverSigningCheck{osInfo},
+func getChecks(osInfo *osinfo.OSInfo) []precheck.Check {
+	return []precheck.Check{
+		&precheck.OSVersionCheck{OSInfo: osInfo},
+		&precheck.DisksCheck{},
+		&precheck.SSHCheck{},
+		&precheck.PowershellCheck{},
+		&precheck.SHA2DriverSigningCheck{OSInfo: osInfo},
 	}
 }
 
@@ -49,7 +53,7 @@ func main() {
 	log = logger.Init("Precheck", false, false, mw)
 	defer log.Close()
 
-	if err = checkRoot(); err != nil {
+	if err = precheck.CheckRoot(); err != nil {
 		logger.Fatal(err)
 	}
 
@@ -62,11 +66,11 @@ func main() {
 	wg := sync.WaitGroup{}
 	for _, c := range checks {
 		wg.Add(1)
-		go func(c check) {
+		go func(c precheck.Check) {
 			defer wg.Done()
-			report, err := c.run()
+			report, err := c.Run()
 			if err != nil {
-				log.Errorf("%s error: %v", c.getName(), err)
+				log.Errorf("%s error: %v", c.GetName(), err)
 			} else {
 				fmt.Println(report.String())
 			}
