+++ cli_tools/import_precheck/precheck/check.go
@@ -1,0 +1,24 @@
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package precheck
+
+// Check represents a single precheck step.
+type Check interface {
+	// GetName returns the name of the precheck step; this is shown to the user.
+	GetName() string
+
+	// Run executes the precheck step.
+	Run() (*Report, error)
+}
