+++ cli_tools/import_precheck/precheck/check_disks_linux_test.go
@@ -1,4 +1,18 @@
-package main
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package precheck
 
 import (
 	"io/ioutil"
@@ -46,7 +60,7 @@ func Test_readMounts(t *testing.T) {
 		},
 	} {
 		t.Run(tc.lsblkOutputFile, func(t *testing.T) {
-			actual, err := (&disksCheck{
+			actual, err := (&DisksCheck{
 				lsblkOverride: func() ([]byte, error) {
 					return ioutil.ReadFile("testdata/" + tc.lsblkOutputFile)
 				},
@@ -97,7 +111,7 @@ func Test_run(t *testing.T) {
 		},
 	} {
 		t.Run(tc.lsblkOutputFile, func(t *testing.T) {
-			report, err := (&disksCheck{
+			report, err := (&DisksCheck{
 				lsblkOverride: func() ([]byte, error) {
 					return ioutil.ReadFile("testdata/" + tc.lsblkOutputFile)
 				},
@@ -108,7 +122,7 @@ func Test_run(t *testing.T) {
 					bytes[511] = 0xAA
 					return bytes, nil
 				},
-			}).run()
+			}).Run()
 			if err != nil {
 				t.Fatal(err)
 			}
