+++ cli_tools/import_precheck/precheck/check_disks_windows.go
@@ -1,16 +1,18 @@
-/*
-Copyright 2017 Google Inc. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package main
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package precheck
 
 import (
 	"encoding/binary"
@@ -21,14 +23,16 @@ import (
 	"golang.org/x/sys/windows"
 )
 
-type disksCheck struct{}
+type DisksCheck struct{}
 
-func (c *disksCheck) getName() string {
+// GetName returns the name of the precheck step; this is shown to the user.
+func (c *DisksCheck) GetName() string {
 	return "Disks Check"
 }
 
-func (c *disksCheck) run() (*report, error) {
-	r := &report{name: c.getName()}
+// Run executes the precheck step.
+func (c *DisksCheck) Run() (*Report, error) {
+	r := &Report{name: c.GetName()}
 
 	sysRoot := os.Getenv("SYSTEMROOT")
 	rootDrive := sysRoot[:2]
