+++ cli_tools/import_precheck/precheck/check_osversion.go
@@ -1,16 +1,18 @@
-/*
-Copyright 2017 Google Inc. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-package main
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package precheck
 
 import (
 	"fmt"
@@ -26,18 +28,21 @@ const (
 	docsURL = "https://cloud.google.com/sdk/gcloud/reference/compute/images/import"
 )
 
-type osVersionCheck struct {
-	osInfo *osinfo.OSInfo
+// OSVersionCheck is a precheck.Check that verifies the disk's operating system is importable.
+type OSVersionCheck struct {
+	OSInfo *osinfo.OSInfo
 }
 
-func (c *osVersionCheck) getName() string {
+// GetName returns the name of the precheck step; this is shown to the user.
+func (c *OSVersionCheck) GetName() string {
 	return "OS Version Check"
 }
 
-func (c *osVersionCheck) run() (*report, error) {
-	r := &report{name: c.getName()}
+// Run executes the precheck step.
+func (c *OSVersionCheck) Run() (*Report, error) {
+	r := &Report{name: c.GetName()}
 	// Find osID from OS config's detection results.
-	major, minor := splitOSVersion(c.osInfo.Version)
+	major, minor := splitOSVersion(c.OSInfo.Version)
 	osID := c.createOSID(major, minor, r)
 	if osID == "" {
 		r.Info("Unable to determine whether your system is supported for import. " +
@@ -55,10 +60,10 @@ func (c *osVersionCheck) run() (*report, error) {
 		}
 	}
 	if supported {
-		if c.osInfo.ShortName == osinfo.Windows {
+		if c.OSInfo.ShortName == osinfo.Windows {
 			// Emit the NT version for Windows, since the same NT version is
 			// either Desktop or Server, and we don't want to emit a misleading message.
-			r.Info(fmt.Sprintf("Detected Windows version number: NT %s", c.osInfo.Version))
+			r.Info(fmt.Sprintf("Detected Windows version number: NT %s", c.OSInfo.Version))
 		} else {
 			r.Info(fmt.Sprintf("Detected system: %s", osID))
 		}
@@ -70,10 +75,10 @@ func (c *osVersionCheck) run() (*report, error) {
 
 // createOSID creates the osID, as used in the `--os` flag of the CLI tools. An empty string is
 // return when unable to determine the osID.
-func (c *osVersionCheck) createOSID(originalMajor string, originalMinor string, r *report) string {
+func (c *OSVersionCheck) createOSID(originalMajor string, originalMinor string, r *Report) string {
 	major, minor := originalMajor, originalMinor
 
-	switch c.osInfo.ShortName {
+	switch c.OSInfo.ShortName {
 	case "":
 		r.Info("Unable to determine OS.")
 		return ""
@@ -91,7 +96,7 @@ func (c *osVersionCheck) createOSID(originalMajor string, originalMinor string, 
 		}
 	}
 
-	release, err := distro.FromComponents(c.osInfo.ShortName, major, minor, c.osInfo.Architecture)
+	release, err := distro.FromComponents(c.OSInfo.ShortName, major, minor, c.OSInfo.Architecture)
 	if err != nil {
 		r.Info(err.Error())
 		return ""
@@ -102,8 +107,8 @@ func (c *osVersionCheck) createOSID(originalMajor string, originalMinor string, 
 	}
 	// If the distro package can't determine the osID, attempt to create one using
 	// the format "os-version".
-	if c.osInfo.ShortName != osinfo.Linux && c.osInfo.ShortName != "" && c.osInfo.Version != "" {
-		return fmt.Sprintf("%s-%s", c.osInfo.ShortName, c.osInfo.Version)
+	if c.OSInfo.ShortName != osinfo.Linux && c.OSInfo.ShortName != "" && c.OSInfo.Version != "" {
+		return fmt.Sprintf("%s-%s", c.OSInfo.ShortName, c.OSInfo.Version)
 	}
 	return ""
 }
