+++ cli_tools/import_precheck/precheck/testdata/lsblk-efi.json
