+++ cli_tools/import_precheck/precheck/testdata/lsblk-one-partition.json
