+++ cli_tools/import_precheck/precheck/testdata/lsblk-single-disk-lvm.json
