+++ cli_tools/import_precheck/precheck/user_linux.go
@@ -1,0 +1,29 @@
+//  Copyright 2017 Google Inc. All Rights Reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+package precheck
+
+import (
+	"errors"
+	"os"
+)
+
+// CheckRoot returns an error if the process's runtime user isn't root.
+func CheckRoot() error {
+	uid := os.Geteuid()
+	if uid != 0 {
+		return errors.New("must be run as root")
+	}
+	return nil
+}
